Add RealmTalksType
