+++ PyConJP2016/DataSource/Timeline/TimelineDataSource.swift
@@ -1,35 +1,18 @@
 //
-//  BookmarkListDataSource.swift
+//  TimelineDataSource.swift
 //  PyConJP2016
 //
-//  Created by Yutaro Muta on 2016/08/18.
+//  Created by Yutaro Muta on 8/19/16.
 //  Copyright © 2016 PyCon JP. All rights reserved.
 //
 
 import UIKit
-import RealmSwift
 
-class BookmarkListDataSource: NSObject, UITableViewDataSource, TimelineDataSource {
+class TimelineDataSource: NSObject, UITableViewDataSource {
+    
+    let reuseIdentifier = "TalkTableViewCell"
     
     var timelines: [Timeline] = []
-    
-    let filterPredicate = NSPredicate(format: "favorited == %@", true)
-    let sortProperties = [SortDescriptor(property: "date", ascending: true), SortDescriptor(property: "place", ascending: true)]
-    
-    func refreshData() {
-        timelines.removeAll()
-        loadTalkObjects({ result in
-            switch result {
-            case .Success(let talks):
-                let keys = talks.map { $0.day }.unique()
-                for tuple in keys.enumerate() {
-                    self.timelines.append(Timeline(key: keys[tuple.index], talks: talks.filter { $0.day == keys[tuple.index]}))
-                }
-                break
-            case .Failure: break
-            }
-        })
-    }
     
     // MARK: - Table View Controller Data Source
